requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.2.0
    # via kombu
anyio==4.6.2.post1
    # via watchfiles
asgiref==3.8.1
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
astroid==2.15.8
    # via
    #   pylint
    #   pylint-celery
async-timeout==4.0.3
    # via redis
attrs==24.2.0
    # via
    #   service-identity
    #   twisted
autobahn==24.4.2
    # via daphne
automat==24.8.1
    # via twisted
bcrypt==4.1.2
    # via -r requirements.in
billiard==3.6.4.0
    # via celery
black==24.3.0
    # via -r requirements.in
boto3==1.34.67
    # via -r requirements.in
botocore==1.34.162
    # via
    #   boto3
    #   s3transfer
build==1.2.2.post1
    # via pip-tools
celery==5.2.6
    # via -r requirements.in
certifi==2024.8.30
    # via
    #   cloudinary
    #   requests
    #   sentry-sdk
cffi==1.17.1
    # via cryptography
cfgv==3.4.0
    # via pre-commit
channels==4.1.0
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==4.2.0
    # via -r requirements.in
charset-normalizer==2.1.1
    # via requests
click==8.1.7
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cloudinary==1.41.0
    # via -r requirements.in
constantly==23.10.4
    # via twisted
coverage==7.4.4
    # via -r requirements.in
cryptography==43.0.3
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==4.1.2
    # via -r requirements.in
decorator==4.4.2
    # via plivo
dill==0.3.9
    # via pylint
distlib==0.3.9
    # via virtualenv
django==5.1.1
    # via
    #   -r requirements.in
    #   channels
    #   django-extensions
    #   django-filter
    #   django-modeltranslation
    #   django-phonenumber-field
    #   django-storages
    #   djangorestframework
    #   easy-thumbnails
django-extensions==3.2.0
    # via -r requirements.in
django-filter==24.1
    # via -r requirements.in
django-ipware==7.0.1
    # via -r requirements.in
django-light==0.1.0.post3
    # via -r requirements.in
django-modeltranslation==0.18.11
    # via -r requirements.in
django-phonenumber-field[phonenumberslite]==7.0.1
    # via -r requirements.in
django-storages==1.13.2
    # via -r requirements.in
djangorestframework==3.15.2
    # via -r requirements.in
easy-thumbnails==2.8.5
    # via -r requirements.in
filelock==3.16.1
    # via virtualenv
flake8==4.0.1
    # via
    #   -r requirements.in
    #   flake8-isort
flake8-isort==4.1.1
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
hiredis==2.3.2
    # via -r requirements.in
httptools==0.6.4
    # via uvicorn
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
identify==2.6.1
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   hyperlink
    #   requests
    #   twisted
incremental==24.7.2
    # via twisted
iso3166==2.0.2
    # via -r requirements.in
isort==5.13.2
    # via
    #   flake8-isort
    #   pylint
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.4.2
    # via celery
lazy-object-proxy==1.10.0
    # via astroid
loguru==0.7.2
    # via -r requirements.in
lxml==4.9.4
    # via plivo
mccabe==0.6.1
    # via
    #   flake8
    #   pylint
msgpack==1.1.0
    # via channels-redis
mypy==1.7.1
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
nodeenv==1.9.1
    # via pre-commit
packaging==24.1
    # via
    #   black
    #   build
    #   gunicorn
pathspec==0.12.1
    # via black
phonenumberslite==8.13.48
    # via django-phonenumber-field
pillow==9.3.0
    # via
    #   -r requirements.in
    #   easy-thumbnails
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.6
    # via
    #   black
    #   pylint
    #   virtualenv
plivo==4.40.0
    # via -r requirements.in
pre-commit==2.19.0
    # via -r requirements.in
prompt-toolkit==3.0.48
    # via click-repl
psycopg2==2.9.3
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.1
    # via service-identity
pycodestyle==2.8.0
    # via flake8
pycparser==2.22
    # via cffi
pyflakes==2.4.0
    # via flake8
pyjwt==2.9.0
    # via plivo
pylint==2.17.7
    # via
    #   pylint-celery
    #   pylint-django
    #   pylint-plugin-utils
pylint-celery==0.3
    # via -r requirements.in
pylint-django==2.5.3
    # via -r requirements.in
pylint-plugin-utils==0.8.2
    # via
    #   pylint-celery
    #   pylint-django
pyopenssl==24.2.1
    # via twisted
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==0.20.0
    # via
    #   -r requirements.in
    #   uvicorn
python-ipware==3.0.0
    # via django-ipware
python-magic==0.4.27
    # via -r requirements.in
pytz==2022.1
    # via
    #   -r requirements.in
    #   celery
pyyaml==6.0.2
    # via
    #   pre-commit
    #   uvicorn
qrcode==8.0
    # via -r requirements.in
redis==5.0.3
    # via
    #   -r requirements.in
    #   channels-redis
requests==2.28.1
    # via
    #   -r requirements.in
    #   plivo
s3transfer==0.10.3
    # via boto3
sentry-sdk==1.44.0
    # via -r requirements.in
service-identity==24.2.0
    # via twisted
six==1.16.0
    # via
    #   cloudinary
    #   plivo
    #   python-dateutil
sniffio==1.3.1
    # via anyio
sqlparse==0.5.1
    # via django
testfixtures==6.18.5
    # via flake8-isort
toml==0.10.2
    # via pre-commit
tomlkit==0.13.2
    # via pylint
twisted[tls]==24.7.0
    # via daphne
txaio==23.1.1
    # via autobahn
typing-extensions==4.12.2
    # via
    #   django-modeltranslation
    #   mypy
    #   twisted
tzdata==2024.2
    # via kombu
urllib3==1.26.20
    # via
    #   botocore
    #   cloudinary
    #   requests
    #   sentry-sdk
uvicorn[standard]==0.32.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
uwsgi==2.0.21
    # via -r requirements.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
virtualenv==20.27.0
    # via pre-commit
wand==0.6.11
    # via -r requirements.in
watchfiles==0.24.0
    # via uvicorn
wcwidth==0.2.13
    # via prompt-toolkit
websockets==13.1
    # via uvicorn
wheel==0.44.0
    # via pip-tools
whitenoise==6.7.0
    # via -r requirements.in
wrapt==1.16.0
    # via astroid
zope-interface==7.1.1
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
